+++ src/plone/app/folder/nogopip.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+from zope.deferredimport import deprecated
+
+
+deprecated(
+    "Please import from plone.folder.nogopip",
+    GopipIndex='plone.folder.nogopip:GopipIndex',
+)